Added missing static_cast to facilitate working with non-built-in types.

+++ computerVision/brick/computerVision/cameraIntrinsicsRational_impl.hh
@@ -398,7 +398,7 @@ namespace brick {
       // The if clause here is to avoid Vector2D throwing an exception
       // if point.z() == 0.
       brick::numeric::Vector2D<FloatType> normalizedPoint;
-      if(point.z() != 0.0) {
+      if(point.z() != static_cast<FloatType>(0.0)) {
         normalizedPoint.setValue(point.x(), point.y(), point.z());
       }
